Refactor utilities for dealing with SoftwareRequirements into own module.
Per great suggestion from Michael on #217.
